requirements.txt
click
google-cloud-storage
pip
protobuf~=5.29.4
pyyaml
requests~=2.32.3
shippo>=3.0.0
six
wrapt

fastapi~=0.115.12
pydantic-settings~=2.9.1
pydantic~=2.11.4
httpx~=0.28.1
uvicorn~=0.34.2
starlette~=0.46.2
algoliasearch~=4.17.0
pg8000~=1.31.2
asyncpg~=0.30.0
stripe~=12.1.0